Prevent comment form from refreshing on submission
The comment form has an id of `#new_comment`. We can use this to bind
to the form’s submit event. The code inside our function will run when
the `submit` event is triggered.

The first thing that we will want to do is prevent the form from
actually submitting, which would cause the page to refresh.