Adding a comment to the feed - part 1
Our end goal is to be able to add our new comment to the page without
having to refresh the page. We can start working toward that goal by
creating a String that represents a comment. Then we can find the
`<div>` that contains all of the comments by it’s class
`.comment-feed`. Finally, we can use jQuery’s
[.append()](https://api.jquery.com/append/) to append our comment to
the comment feed.