Create DOM elements for comment
We want our newly added comments to match the style of the other
comments on the page.

The HTML for the other comments looks something like this:

```html
<div class="comment">
  <h4>Eric</h4>
  <p>Hello world</p>
</div>
```

We can use jQuery to create DOM elements as shown in the sample above.
Instead of appending a string to the comment feed, we can append the
DOM element that we just created onto the page.