Create the comment asynchronously
Rather than hard coding the data, we want to use jQuery’s
[.ajax()](https://api.jquery.com/jQuery.ajax/) function to send a POST
request to `/posts/:id/comments` that includes all of the data in our
form. When the request finishes, either the `success` or the `failure`
callback function will run, depending on the [HTTP status
code](http://en.wikipedia.org/wiki/List_of_HTTP_status_codes) of the
response that the server sends back.

Usually our requests to the server are sending back HTML as the body of
the response. In order to configure the server to send back JSON as the
response, you need to change your controller a little bit. This has
already been done for you but you should go check out the
`CommentsController` to see how it works.