Check out the response
Swap our the alert in the `success` callback with a `debugger;` statement.
This works like pry in Ruby does. The only trick is that you need to have the
Javascript Console open when you submit the comment form.

> Hint: You can open the Javascript Console with `cmd + option + J`.

Try creating a comment with the Javascript Console open. Once you're in the
debugging session, type `response` into the Javascript Console to see the value
of the response that was returned from the server.

It looks like our hard coded object!