Add the real comment onto the page :boom:
Since the `response` is a JSON hash that represents the comment created
in the database, we can use this in place of our hard coded comment
attributes object. The easiest way to do this is to move all of the
code responsible for adding the comment to the page into the `success`
callback.